+++ tests/testthat.R
@@ -1,0 +1,3 @@
+library(testthat)
+library(dataraft.adapters)
+test_check("dataraft.adapters")